MGL-114 : Fix medium priority coverity issues
Using invalid iterator (INVALIDATE_ITERATOR)
	modified:   galera/src/certification.cpp
	modified:   gcomm/src/fair_send_queue.hpp

Unchecked return value (CHECKED_RETURN)
	modified:   galera/src/mapped_buffer.cpp
	modified:   galerautils/src/gu_asio_stream_engine.cpp
	modified:   gcomm/src/pc_proto.cpp
	modified:   gcs/src/gcs.cpp
	modified:   gcs/src/gcs_group.cpp

Uncaught exception (UNCAUGHT_EXCEPT)
	modified:   galera/src/mapped_buffer.hpp

Data race condition (MISSING_LOCK)
	modified:   galera/src/replicator_smm.cpp

Copy-paste error (COPY_PASTE_ERROR)
	modified:   galera/src/replicator_str.cpp

Explicit null dereferenced (FORWARD_NULL)
	modified:   galerautils/src/gu_dbug.c
	modified:   galerautils/src/gu_resolver.cpp
	modified:   galerautils/src/gu_to.c
	modified:   gcs/src/gcs_group.cpp

Improper use of negative value (NEGATIVE_RETURNS)
	modified:   galerautils/src/gu_fdesc.cpp
	modified:   gcs/src/gcs.cpp

Unintentional integer overflow (OVERFLOW_BEFORE_WIDEN)
	modified:   galerautils/src/gu_fifo.c

Division or modulo by float zero (DIVIDE_BY_ZERO)
	modified:   galerautils/src/gu_histogram.cpp
	modified:   gcomm/src/evs_proto.cpp

Uninitialized pointer field (UNINIT_CTOR)
	modified:   gcs/src/gcs_recv_msg.hpp

Missing unlock (LOCK)
	modified:   gcs/src/gcs_test.cpp

+++ galera/src/mapped_buffer.hpp
@@ -21,7 +21,7 @@ namespace galera
         typedef gu::byte_t const* const_iterator;
 
         MappedBuffer(const std::string& working_dir, 
-                     size_t threshold = 1 << 20);
+                     size_t threshold = size_t(1) << size_t(20));
 
         ~MappedBuffer();
